Simplify Avatar to just use `user` again
If user.id is specified, but user.image is not and client is, attempt to look up the user via client.api, and get an image asyncrhonously.

Also, adjust layout so the initials show in the gray circle while the image is loading, or if the image fails to load.